partial meld

+++ TDG SQL Tools/sql/functions/tdgMeldAzimuths.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION tdg.tdgMeldBuffers(
+CREATE OR REPLACE FUNCTION tdg.tdgMeldAzimuths(
     target_table_ REGCLASS,
     target_column_ TEXT,
     target_geom_ TEXT,
@@ -6,10 +6,12 @@ CREATE OR REPLACE FUNCTION tdg.tdgMeldBuffers(
     source_column_ TEXT,
     source_geom_ TEXT,
     tolerance_ FLOAT,
+    buffer_geom_ TEXT DEFAULT NULL,
     max_angle_diff_ INTEGER DEFAULT 15,
     line_start_ FLOAT DEFAULT 0.33,
     line_end_ FLOAT DEFAULT 0.67,
-    only_nulls_ BOOLEAN DEFAULT 't'
+    only_nulls_ BOOLEAN DEFAULT 't',
+    nullify_ BOOLEAN DEFAULT 'f'
 )
 RETURNS BOOLEAN AS $func$
 
@@ -17,6 +19,10 @@ DECLARE
     sql TEXT;
     target_srid INTEGER;
     source_srid INTEGER;
+    buffer_srid INTEGER;
+    temp_buffers BOOLEAN;
+    target_pkid TEXT;
+    source_pkid TEXT;
 
 BEGIN
     raise notice 'PROCESSING:';
@@ -28,6 +34,9 @@ BEGIN
     IF NOT tdgColumnCheck(target_table_,target_geom_) THEN
         RAISE EXCEPTION 'Column % not found', target_geom_;
     END IF;
+    IF buffer_geom_ IS NOT NULL AND NOT tdgColumnCheck(target_table_,buffer_geom_) THEN
+        RAISE EXCEPTION 'Column % not found', buffer_geom_;
+    END IF;
     IF NOT tdgColumnCheck(source_table_,source_column_) THEN
         RAISE EXCEPTION 'Column % not found', source_column_;
     END IF;
@@ -38,22 +47,14 @@ BEGIN
     -- srid check
     BEGIN
         RAISE NOTICE 'Getting SRID of target geometry';
-        EXECUTE 'SELECT tdgGetSRID($1,$2);'
-        USING   target_table_,
-                target_geom_
-        INTO    target_srid;
-
+        target_srid := tdgGetSRID(target_table_,target_geom_);
         IF target_srid IS NULL THEN
             RAISE EXCEPTION 'ERROR: Cannot determine the srid of the geometry in table %', target_table_;
         END IF;
         raise NOTICE '  -----> SRID found %',target_srid;
 
         RAISE NOTICE 'Getting SRID of source geometry';
-        EXECUTE 'SELECT tdgGetSRID($1,$2);'
-        USING   source_table_,
-                source_geom_
-        INTO    source_srid;
-
+        source_srid := tdgGetSRID(source_table_,source_geom_);
         IF source_srid IS NULL THEN
             RAISE EXCEPTION 'ERROR: Cannot determine the srid of the geometry in table %', source_table_;
         END IF;
@@ -62,43 +63,122 @@ BEGIN
         IF NOT target_srid = source_srid THEN
             RAISE EXCEPTION 'SRID on geometry columns do not match';
         END IF;
+
+        IF buffer_geom_ IS NOT NULL THEN
+            RAISE NOTICE 'Getting SRID of buffer geometry';
+            buffer_srid := tdgGetSRID(target_table_,buffer_geom_);
+            IF NOT target_srid = buffer_srid THEN
+                RAISE EXCEPTION 'SRID on geometry columns do not match';
+            END IF;
+        END IF;
     END;
 
+    -- get target and source pkid columns
+    target_pkid := tdg.tdgGetPkColumn(target_table_);
+    source_pkid := tdg.tdgGetPkColumn(source_table_);
+
     BEGIN
+        -- set nulls
+        IF nullify_ THEN
+            EXECUTE '
+                UPDATE  '||target_table_||'
+                SET     '||target_column_||' = NULL';
+        END IF;
+
+        -- add buffer geom column
+        IF buffer_geom_ IS NULL THEN
+            temp_buffers := 't';
+            RAISE NOTICE 'Buffering...';
+            EXECUTE '
+                ALTER TABLE '||target_table_||'
+                ADD COLUMN  tmp_buffer_geom geometry(multipolygon,'||target_srid::TEXT||')';
+
+            sql := '
+                UPDATE  '||target_table_||'
+                SET     tmp_buffer_geom = ST_Multi(
+                            ST_Buffer(
+                                '||target_geom_||',
+                                '||tolerance_::TEXT||',
+                                ''endcap=flat''
+                            )
+                        )';
+            IF only_nulls_ THEN
+                EXECUTE sql || ' WHERE '||target_column_||' IS NULL';
+            ELSE
+                EXECUTE sql;
+            END IF;
+
+            -- add buffer index
+            RAISE NOTICE 'Indexing buffer...';
+            EXECUTE '
+                CREATE INDEX tsidx_meldgeom
+                ON '||target_table_||'
+                USING GIST (tmp_buffer_geom)';
+            EXECUTE 'ANALYZE '||target_table_||' (tmp_buffer_geom)';
+
+            buffer_geom_ := 'tmp_buffer_geom';
+        ELSE
+            temp_buffers := 'f';
+        END IF;
+
         -- check for matches
         RAISE NOTICE 'Getting azimuth matches';
         sql := '
+            WITH target_azi AS (
+                SELECT  '||target_pkid' AS id,
+                        ST_Azimuth(
+                            ST_LineInterpolatePoint(
+                                src.'||target_geom_||',
+                                '||line_start_::TEXT||'
+                            ),
+                            ST_LineInterpolatePoint(
+                                src.'||target_geom_||',
+                                '||line_end_::TEXT||'
+                            )
+                        )
+            ),
+            source_azi AS (
+                SELECT  '||source_pkid||' AS id
+                        ST_Azimuth(
+                            ST_LineInterpolatePoint(
+                                src.'||source_geom_||',
+                                '||line_start_::TEXT||'
+                            ),
+                            ST_LineInterpolatePoint(
+                                src.'||source_geom_||',
+                                '||line_end_::TEXT||'
+                            )
+                        )
+            )
             UPDATE  '||target_table_||'
             SET     '||target_column_||' = (
                         SELECT      src.'||source_column_||'
-                        FROM        '||source_table_||' src
-                        WHERE       ST_DWithin(
-                                        '||target_table_||'.'||target_geom_||',
-                                        src.'||source_geom_||',
-                                        '||tolerance_::TEXT||'
-                                    )
+                        FROM        '||source_table_||' src,
+                                    target_azi,
+                                    source_azi
+                        WHERE       '||target_table_||'.'||target_pkid||' = target_azi.id
+                        AND         '||source_table_||'.'||source_pkid||' = source_azi.id
                         AND         ST_Intersects(
-                                        ST_Buffer(
-                                            '||target_table_||'.'||target_geom_||',
-                                            'tolerance_::TEXT'
-                                        ),
+                                        '||target_table_||'.'||buffer_geom_||',
                                         src.'||source_geom_||'
                                     )
-                        AND         ST_Length(
-                                        ST_Intersection(
-                                            tmp_buffer_geom,
-                                            src.'||source_geom_||'
-                                        )
-                                    ) >= '||min_shared_length_pct_::FLOAT||' * ST_Length('||target_table_||'.'||target_geom_||')
-                        ORDER BY    ST_Length(
-                                        ST_Intersection(
-                                            tmp_buffer_geom,
-                                            src.'||source_geom_||'
-                                        )
-                                    ) DESC
+                        AND         '!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!'
+                        ORDER BY    ST_Distance(
+                                        ST_LineInterpolatePoint(
+                                            '||target_table_||'.'||target_geom_||',
+                                            '||line_start_::TEXT||'
+                                        ),
+                                        src.'||source_geom_||'
+                                    ) + ST_Distance(
+                                        ST_LineInterpolatePoint(
+                                            '||target_table_||'.'||target_geom_||',
+                                            '||line_end_::TEXT||'
+                                        ),
+                                        src.'||source_geom_||'
+                                    ) / 2
+                                    ASC
                         LIMIT       1
-                    )
-            WHERE   ST_Length('||target_table_||'.'||target_geom_||') > '||min_target_length_::TEXT;
+                    )';
 
         IF only_nulls_ THEN
             EXECUTE sql || ' AND '||target_table_||'.'||target_column_||' IS NULL';
@@ -106,12 +186,14 @@ BEGIN
             EXECUTE sql;
         END IF;
 
-        -- drop temporary buffers
-        RAISE NOTICE 'Dropping temporary buffers';
-        EXECUTE 'ALTER TABLE '||target_table_||' DROP COLUMN tmp_buffer_geom';
+        IF temp_buffers THEN
+            -- drop temporary buffers
+            RAISE NOTICE 'Dropping temporary buffers';
+            EXECUTE 'ALTER TABLE '||target_table_||' DROP COLUMN tmp_buffer_geom';
+        END IF;
     END;
 
     RETURN 't';
 END $func$ LANGUAGE plpgsql;
-ALTER FUNCTION tdg.tdgMeldBuffers(REGCLASS,TEXT,TEXT,REGCLASS,TEXT,TEXT,FLOAT,
-    BOOLEAN,FLOAT,FLOAT) OWNER TO gis;
+ALTER FUNCTION tdg.tdgMeldAzimuths(REGCLASS,TEXT,TEXT,REGCLASS,TEXT,TEXT,FLOAT,
+    TEXT,INTEGER,FLOAT,FLOAT,BOOLEAN,BOOLEAN) OWNER TO gis;
